Updates
1- created the paragraph limit
2- moved the delete document function to management
3- updated the scrape document template
