+++ django_backend/chatbot_be/models/scraped_data.py
@@ -7,6 +7,7 @@ class ScrapedData(models.Model):
     binary_content = models.BinaryField(null=True, blank=True)  # For binary content
     pdf_file = models.FileField(upload_to='uploads/', null=True, blank=True)
     created_at = models.DateTimeField(auto_now_add=True)
+    title = models.CharField(max_length=100)
 
     def __str__(self):
         return f'Scraped from {self.url} ({self.file_type})'
@@ -17,7 +18,8 @@ class ScrapedDataMeta(models.Model):
     file_type = models.CharField(max_length=50)
     pdf_file = models.FileField(upload_to='uploads/', null=True, blank=True)
     created_at = models.DateTimeField()
-    content_preview = models.CharField(max_length=100, blank=True, null=True)  # Store a short preview
-
+    content_preview = models.CharField(max_length=100, blank=True, null=True) # Store a short preview
+    title = models.CharField(max_length=100)
+    
     def __str__(self):
         return f'Metadata for {self.url} ({self.file_type})'
