+++ django_backend/chatbot_be/templates/base.html
@@ -28,7 +28,7 @@
         {% endblock %}
     </div>
 
-    <div class="footer">
+    <div class="footer", style="position: fixed; bottom: 0;">
         Copyright 2025. All rights reserved
     </div>
