+++ django_backend/chatbot_be/templates/processor_url.html
@@ -1,13 +1,13 @@
 <!-- URL Input Section -->
 <div class="mb-3">
     <label for="url" class="form-label">URL:</label>
-    <input type="text" class="form-control" id="url" placeholder="Enter the URL to scrape" required disabled>
+    <input type="text" class="form-control" id="url" placeholder="Enter the URL to scrape" required>
 </div>
 
 <!-- Output File Name Section -->
 <div class="mb-3">
     <label for="output_file" class="form-label">Output File Name:</label>
-    <input type="text" class="form-control" id="output_file" placeholder="Enter desired output file name" required disabled>
+    <input type="text" class="form-control" id="output_file" placeholder="Enter desired output file name" required>
 </div>
 
 <!-- Scrape Button -->
